Override existing values from core, do not substitute top-level keys
This fixes an issue by which if a upper layer (theme, user) partially
set the design tokens the array won't contain the other parts.

If a theme defines only one color (ex: primary) we still want
the other ones in the color array (background, text),
just with the values from core.