Don't show title permalink editor until a save
This somewhat diminishes the effects of this pull request, but it will
help it pass the e2e tests and also streamlines writing a bit. Reverts
back to preventing showing the permalink editor in the title until after
a manual or autosave. It does at least show the actual permalink on
the autosave and users can edit the slug at any time using the
document sidebar.